pyproject.toml
[project]
name = "bestwishes"
version = "0.1.0"
description = ""
authors = [
    {name = "funnymilk",email = "andre.tum91@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django (>=4.2.0,<5.0.0)",
    "djangorestframework (>=3.15.0,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.5.2,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
